Actions and Processes
Main file starts the ImageProcessor which builds a list of actions to be performed.
Actions are an enum encapsulated with a class where each action is associated with a function.
Actions are built using a separate factory. 
High level ImageProcessor generates the list of actions (this will be from a yaml file) and processes them in turn. 
CMake configuration for this project is included