OpenCV integration
i have integrated OpenCV into the project and used it to load, convert to greyscale and save an image.
The process action function takes in a pointer to an image which can be manipulated at each step. 
To do so I have had to replace histogram with greyscale because of the complexity surrounding histogram. 
The output location however is not currently working.