Image struct
Images are decoupled from action's so that an image contains only the data which is passed to and returned from each action.
A test image has been committed which is used to test an image loads and is copied to the build system for running using ctest.
Readme has been updated.
Temporarily a new cmake target action_new has been added which shall replace the old action target when the conversion has been completed.