Action's class and Subclasses
Actions now follow a standarised pattern of being initialised by the yaml file and applied to an image when processed by defining the process().
Each subclass represents a single image.
Saving now hold's the counter for adding a suffix when a name sin't supplied and the associated Image test has been removed.