DI attempt through file system interface
Using Dependency Injection to wrap calls to the file system.
A mock of the interface are part of the test framework to enable easier testing of process() member functions. The interface is scoped to the life of the Action objects.
Raw pointers are used because of the additional difficulty of the unique_ptr operations during tests.
Each action now takes a pointer to the Filesystem object where necessary (Initialise and Save) with file system calls replaced by a call to Filesystem member functions. Some std::move for unqiue_ptrs have remained.
Removed erroneous double call of the first action process() call from image_processor.cpp