OpenCV functions wrapper
To maintain Dependency Injection (DI) with OpenCV a wrapper class for OpenCV functions (non-file system calls) has been created in the same vein as Filsystem().
This includes updates to the CMake variables, modification to a general purpose DI file, and where called necessary the IOpenCV class is now integrated with a test written for Grey.
Other changes include fixing the equality test for images to work as expected and be called in the appropriate tests. The Initialise::process test had a bug due to an inconsistent name being returned from the im1_.output_folder_ and output.output_folder_. By setting im1_.output_folder_ to be "output" (the default) the bug is fixed.