Action Factory Tests
Action Factory's create of actions now has tests which are tested against any available variables stored during creation as well as a new variable assigned to the base class.
For each derived action any and all values set by constructor can now be called (which is used to test if the Action was created properly).
Improved ParseMap operator "[]" which now returns an error from the application to make it easier to track down bugs due to it and added the keyword return to returning size().
Minor formatting fixes.